+++ src/Components/Redux/action.jsx
@@ -5,6 +5,7 @@ export const requestCars = () => async (dispatch) => {
   dispatch({
     type: CAR.LOAD
   })
+
   try {
     const json = await axios.get('MOCK_DATA.json')
     console.log(json)
@@ -25,6 +26,13 @@ export const requestCars = () => async (dispatch) => {
 export const deleteCar = (car) => async (dispatch) => {
   dispatch({
     type: CAR.DELETE_CAR,
+    car
+  })
+}
+
+export const updateCar = (car) => async (dispatch) => {
+  dispatch({
+    type: CAR.UPDATE_CAR,
     car
   })
 }
